+++ rust/src/intcode/io.rs
@@ -5,7 +5,7 @@ pub trait IntcodeInput {
 }
 
 pub trait IntcodeOutput {
-    fn write(&mut self, out: isize);
+    fn write(&mut self, out: i64);
 }
 
 impl IntcodeInput for std::io::Empty {
@@ -26,7 +26,7 @@ impl IntcodeInput for &str {
     }
 }
 
-impl IntcodeInput for Vec<isize> {
+impl IntcodeInput for Vec<i64> {
     fn read_line(&mut self) -> Option<String> {
         if self.len() == 0 {
             return None;
@@ -35,31 +35,31 @@ impl IntcodeInput for Vec<isize> {
     }
 }
 
-impl IntcodeInput for Option<isize> {
+impl IntcodeInput for Option<i64> {
     fn read_line(&mut self) -> Option<String> {
         self.take().map(|i| i.to_string())
     }
 }
 
 impl IntcodeOutput for std::io::Sink {
-    fn write(&mut self, _: isize) {}
+    fn write(&mut self, _: i64) {}
 }
 impl IntcodeOutput for std::io::StdoutLock<'_> {
-    fn write(&mut self, out: isize) {
+    fn write(&mut self, out: i64) {
         match writeln!(self, "OUTPUT: {}", out) {
             // silently fail
             _ => (),
         };
     }
 }
-impl IntcodeOutput for Vec<isize> {
-    fn write(&mut self, out: isize) {
+impl IntcodeOutput for Vec<i64> {
+    fn write(&mut self, out: i64) {
         self.push(out)
     }
 }
 
-impl IntcodeOutput for Option<isize> {
-    fn write(&mut self, out: isize) {
+impl IntcodeOutput for Option<i64> {
+    fn write(&mut self, out: i64) {
         *self = Some(out);
     }
 }
